+++ libs/utils/src/lib/image-uploader.ts
@@ -2,7 +2,21 @@ import { AngularFireStorage } from "@angular/fire/storage";
 import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 import { sanitizeFileName } from "./file-sanitizer";
-import { ImgRef } from "@blockframes/ui/cropper/cropper/cropper.component";
+
+export interface ImgRef {
+  url: string;
+  ref: string;
+  originalRef: string
+}
+
+export function createImgRef(ref: Partial<ImgRef> = {}): ImgRef {
+  return {
+    url: '',
+    ref: '',
+    originalRef: '',
+    ...ref
+  };
+}
 
 @Injectable({ providedIn: 'root' })
 export class ImageUploader {
@@ -17,7 +31,7 @@ export class ImageUploader {
    * @param imageUrl 
    * @param afPath 
    */
-  public async upload(imageUrl: string, afPath: string = 'movies'): Promise<ImgRef> {
+  public async upload(imageUrl: string, afPath: string = 'movies'): Promise<ImgRef|undefined> {
     try {
       const data = await this.httpClient.get(imageUrl, { responseType: 'blob' }).toPromise();
       const snapshot = await this.afStorage.upload(`${afPath}/${sanitizeFileName(imageUrl)}`, data)
@@ -25,7 +39,7 @@ export class ImageUploader {
       const meta = await snapshot.ref.getMetadata();
       return { url, ref: meta.fullPath, originalRef: '' };
     } catch (error) {
-      return;
+      return ;
     }
   }
